+++ libs/@guardian/react-crossword/.storybook/preview.tsx
@@ -1,5 +1,5 @@
 import { FocusStyleManager, palette } from '@guardian/source/foundations';
-import type { Decorator, Preview } from '@storybook/react';
+import type { Decorator, Preview } from '@storybook/react-vite';
 import { useEffect } from 'react';
 import { viewport } from './preview/viewport';
 
@@ -26,31 +26,40 @@ const preview: Preview = {
 			},
 		},
 		backgrounds: {
-			default: 'palette.neutral[100]',
-			values: [
-				{
+			options: {
+				'palette.neutral[100]': {
 					name: 'palette.neutral[100]',
 					value: palette.neutral[100],
 				},
-				{
+
+				'palette.neutral[97]': {
 					name: 'palette.neutral[97]',
 					value: palette.neutral[97],
 				},
-				{
+
+				'palette.neutral[10]': {
 					name: 'palette.neutral[10]',
 					value: palette.neutral[10],
 				},
-				{
+
+				'palette.brand[400]': {
 					name: 'palette.brand[400]',
 					value: palette.brand[400],
 				},
-				{
+
+				'palette.brandAlt[400]': {
 					name: 'palette.brandAlt[400]',
 					value: palette.brandAlt[400],
 				},
-			],
+			},
 		},
 		viewport,
+	},
+
+	initialGlobals: {
+		backgrounds: {
+			value: 'palette.neutral[100]',
+		},
 	},
 };