+++ libs/@guardian/react-crossword/src/components/Cell.stories.tsx
@@ -1,4 +1,4 @@
-import type { Meta, StoryObj } from '@storybook/react';
+import type { Meta, StoryObj } from '@storybook/react-vite';
 import { groupedClues as data } from '../../stories/formats/grouped-clues';
 import type { Theme } from '../@types/crossword';
 import { ContextProvider } from '../context/ContextProvider';