+++ libs/@guardian/react-crossword/src/components/Grid.tsx
@@ -293,7 +293,7 @@ export const Grid = () => {
 					deleteLetter(event.target.value);
 				}
 			} else {
-				if (event.key && event.key.length === 1) {
+				if (event.key.length === 1) {
 					event.preventDefault();
 					typeLetter(event.key);
 				}