+++ libs/@guardian/source-development-kitchen/src/react-components/star-rating/StarRating.stories.tsx
@@ -1,4 +1,4 @@
-import type { Meta, StoryFn } from '@storybook/react';
+import type { Meta, StoryFn } from '@storybook/react-vite';
 import type { StarRatingProps } from './StarRating';
 import { StarRating } from './StarRating';