+++ libs/@guardian/source/scripts/create-icons/get-svgs-from-figma.ts
@@ -49,7 +49,7 @@ export const getIconsFromFigma = async () => {
 	const figmaIconComponents = figmaComponents.filter((c) => {
 		return (
 			// Only get icons that are in a certain frame
-			c.containing_frame && c.containing_frame.name === 'UI icons 24x24'
+			c.containing_frame?.name === 'UI icons 24x24'
 		);
 	});