+++ libs/@guardian/source/src/react-components/container/Container.stories.tsx
@@ -1,4 +1,4 @@
-import type { Meta, StoryObj } from '@storybook/react';
+import type { Meta, StoryObj } from '@storybook/react-vite';
 import { Container } from './Container';
 
 const meta: Meta<typeof Container> = {