+++ libs/@guardian/source/src/react-components/text-input/TextInput.stories.tsx
@@ -1,4 +1,4 @@
-import type { Meta, StoryObj } from '@storybook/react';
+import type { Meta, StoryObj } from '@storybook/react-vite';
 import { useState } from 'react';
 import { palette } from '../../foundations';
 import { TextInput } from './TextInput';
@@ -32,7 +32,6 @@ type Story = StoryObj<typeof TextInput>;
 
 const Template: Story = {
 	render: (args) => {
-		// eslint-disable-next-line react-hooks/rules-of-hooks -- it _is_ actually a react function component
 		const [state, setState] = useState('');
 		return (
 			<TextInput
@@ -243,9 +242,9 @@ export const SupportingTextCustomTheme: Story = {
 			backgroundInput: palette.neutral[20],
 		},
 	},
-	parameters: {
+	globals: {
 		backgrounds: {
-			default: 'palette.neutral[10]',
+			value: 'palette.neutral[10]',
 		},
 	},
 };
@@ -262,9 +261,9 @@ export const ErrorWithMessageCustomTheme: Story = {
 			backgroundInput: palette.neutral[20],
 		},
 	},
-	parameters: {
+	globals: {
 		backgrounds: {
-			default: 'palette.neutral[10]',
+			value: 'palette.neutral[10]',
 		},
 	},
 };