+++ libs/@guardian/ab-core/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @guardian/ab-core
 
+## 9.0.0
+
+### Major Changes
+
+- 816eebd:
+  - Peer Dependency: Update `typescript` to `5.9.3`
+  - Peer Dependency: Update `tslib` to `2.8.1`
+
+## 8.0.2
+
+### Patch Changes
+
+- 56dab8a: no-op
+
+  This is a no-op release to test migration to [NPM trusted publishing](https://docs.npmjs.com/trusted-publishers).
+
 ## 8.0.1
 
 ### Patch Changes
@@ -49,7 +65,6 @@
   In order to update your code to work identically, you must now provide the
   following keys to the constructor’s argument. They are listed here along with
   the fallbacks previously applied:
-
   - `serverSideTests` &rarr; `{}`
   - `errorReporter` &rarr; `() => undefined`
   - `ophanRecord` &rarr; `() => undefined`