+++ libs/@guardian/newsletter-types/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @guardian/newsletter-types
 
+## 4.0.0
+
+### Major Changes
+
+- 816eebd:
+  - Peer Dependency: Update `typescript` to `5.9.3`
+  - Peer Dependency: Update `tslib` to `2.8.1`
+
+## 3.0.2
+
+### Patch Changes
+
+- 56dab8a: no-op
+
+  This is a no-op release to test migration to [NPM trusted publishing](https://docs.npmjs.com/trusted-publishers).
+
 ## 3.0.1
 
 ### Patch Changes